fix: move components to auth folder
